+++ thrift/src/test/scala/com/twitter/finatra/thrift/tests/ThriftWarmupTest.scala
@@ -5,12 +5,10 @@ import com.twitter.finatra.thrift.routing.{ThriftWarmup, ThriftRouter}
 import com.twitter.inject.Test
 
 class ThriftWarmupTest extends Test {
-
   test("ThriftWarmup refuses to send a request if a router is not configured") {
     intercept[IllegalStateException] {
       val tw = new ThriftWarmup(new ThriftRouter(null, null))
       tw.send(Uppercase, Uppercase.Args("hithere"), 1)()
     }
   }
-
 }